+++ communitylighting.js
@@ -13,8 +13,8 @@ class CommunityLighting {
         "communitylighting-parseauthors": (obj) => {
             var authors = Object.keys(obj);
             var data = "";
-            authors.forEach((author)=>{
-                data+=`<p>${author}</p>
+            authors.forEach((author) => {
+                data += `<p>${author}</p>
                 `;
             })
             return data;
@@ -25,22 +25,43 @@ class CommunityLighting {
 
         // Get all community lights
         // Add settings to disable specific lights from selection
-    
+
         CommunityLighting.animationManager = new CLAnimationManager();
         await CommunityLighting.animationManager.registerAnimations();
 
         game.settings.registerMenu(CommunityLighting.moduleName, "mySettingsMenu", {
-            name: "COMMUNITYLIGHTING.settings.name",
-            label: "COMMUNITYLIGHTING.settings.label",
+            name: game.i18n.localize("COMMUNITYLIGHTING.settings.name"),
+            label: game.i18n.localize("COMMUNITYLIGHTING.settings.label"),
             icon: "fas fa-lightbulb",
             type: CommunityLightingSettings,
             restricted: true
         });
+        game.settings.register(CommunityLighting.moduleName, 'closeLightOnSubmit', {
+            name: game.i18n.localize('COMMUNITYLIGHTING.settings.closeLightOnSubmit.name'),
+            hint: game.i18n.localize('COMMUNITYLIGHTING.settings.closeLightOnSubmit.hint'),
+            scope: 'client',
+            config: true,
+            type: Boolean,
+            default: true
+        });
+        game.settings.register(CommunityLighting.moduleName, 'closeTokenOnSubmit', {
+            name: game.i18n.localize('COMMUNITYLIGHTING.settings.closeTokenOnSubmit.name'),
+            hint: game.i18n.localize('COMMUNITYLIGHTING.settings.closeTokenOnSubmit.hint'),
+            scope: 'client',
+            config: true,
+            type: Boolean,
+            default: true
+        });
 
         Handlebars.registerHelper(CommunityLighting.handlebarsHelpers);
-        
+
+    }
+
+    static async onReady() {
+        // Patching and using libWrapper if available
+        CLMonkeyPatcher.runPatches()
     }
 }
 
-
 Hooks.on("init", CommunityLighting.onInit);
+Hooks.on("ready", CommunityLighting.onReady);