+++ modules/custompropertymanager.js
@@ -1,85 +1,106 @@
 class CLCustomPropertyManager {
 
-    static async saveCustomProperties(object, changes) {
-        var customProperties = JSON.parse(JSON.stringify(changes.lightAnimation)); // Clone changes
+    static async updateCustomProperties(updateData, lightAnimationChanges) {
+        var customProperties = JSON.parse(JSON.stringify(lightAnimationChanges)); // Clone changes
 
         // Ensure we are not saving the data core stores in the db
-        Object.keys(customProperties).forEach(key=>{
-            if(key == "type" || key == "intensity" || key == "speed"){
+        Object.keys(customProperties).forEach(key => {
+            if (key == "type" || key == "intensity" || key == "speed") {
                 delete customProperties[key];
-            }               
+            }
         });
-        var placeable = canvas.lighting.get(object._id) || canvas.tokens.get(object._id);
-        if(!customProperties || Object.getOwnPropertyNames(customProperties).length == 0){
-            await placeable.unsetFlag("CommunityLighting", "customProperties"); // Remove flag if no custom vars
+
+        let placeableObject = canvas.lighting.get(updateData.id) ?? canvas.tokens.get(updateData.id);
+        if (!customProperties || Object.getOwnPropertyNames(customProperties).length == 0) {
+            // placeableObject.document.unsetFlag("CommunityLighting", "customProperties");
+
+            // No custom properties detected. Previously, we cleared the flags. However, with 0.8,
+            // this is hit when intensity and/or speed only are changed, causing breakage on refresh.
+            console.log("No C-prop")
         } else {
-            await placeable.setFlag("CommunityLighting", "customProperties", customProperties); // Set flag with custom vars
+            //updateData.flags.CommunityLighting = {customProperties: customProperties};
+            placeableObject.document.setFlag("CommunityLighting", "customProperties", customProperties);
         }
-        if(object.actorId){
-            CLCustomPropertyManager.loadCustomProperties(canvas.tokens.get(object._id).light);
+    }
+
+    static async saveCustomProperties(data, flagsChanges) {
+        var customProperties = JSON.parse(JSON.stringify(flagsChanges)); // Clone changes
+
+        if (customProperties) {
+            // Add customProperties from the flag into the lightAnimation object
+            mergeObject(data._source.lightAnimation, customProperties)
+            if (data._source.actor) {
+                data._source.update({lightAnimation: pointSource._source.data.lightAnimation}, {diff: false, loadedProperty: true})
+            }
         }
     }
 
     static loadCustomProperties(pointSource) {
-        if(!pointSource._source){
+        if (!pointSource._source) {
             return; // Source is not cached yet, return
         }
-        var customProperties = pointSource._source.getFlag("CommunityLighting", "customProperties");
+        var customProperties = pointSource.object.getFlag("CommunityLighting", "customProperties");
 
         // Remove any customProperties from the current lightAnimation object
-        Object.keys(pointSource._source.data.lightAnimation).forEach(key=>{
-            if(key != "type" && key != "intensity" && key != "speed"){
-                delete pointSource._source.data.lightAnimation[key];
-            }               
+        Object.keys(pointSource.object.data._source.lightAnimation).forEach(key => {
+            if (key != "type" && key != "intensity" && key != "speed") {
+                delete pointSource.object.data._source.lightAnimation[key];
+            }
         });
 
-        if(customProperties){
+        if (customProperties) {
             // Add customProperties from the flag into the lightAnimation object
-            mergeObject(pointSource._source.data.lightAnimation, customProperties)
-            if(pointSource._source.actor){
-                pointSource._source.update({lightAnimation: pointSource._source.data.lightAnimation}, {diff:false, loadedProperty:true})
+            mergeObject(pointSource.object.data._source.lightAnimation, customProperties)
+            if (pointSource.object.data._source.actor) {
+                pointSource.object.data._source.update({lightAnimation: pointSource._source.lightAnimation}, {diff: false, loadedProperty: true})
             }
         }
 
     }
 
     static async removeAllCustomProperties(html) {
+        html.find(".community-lighting-custom-property-noanim").remove();
         html.find(".community-lighting-custom-property").hide('normal', function () {
             $(this).remove();
         });
     }
 
     static async addCustomProperties(objectConfig, customPropertySibling, customAnimationProperties, animateShow = false) {
+        var customAnimationPropertiesClone = JSON.parse(JSON.stringify(customAnimationProperties)); // Clone object so we can reverse it without reversing original
+        const animation = objectConfig.object.data?.lightAnimation?._source || objectConfig.object.light.animation;
 
-        var isToken = false;
-        if (objectConfig.object.light) {
-            isToken = true;
+        function insertDescription(customPropertyObject){
+            return customPropertyObject.description?`<p class="hint community-lighting-custom-property-noanim">${customPropertyObject.description}</p>`:''
         }
 
-        var customAnimationPropertiesClone = JSON.parse(JSON.stringify(customAnimationProperties)); // Clone object so we can reverse it without reversing original
-        
         customAnimationPropertiesClone.reverse().forEach((customPropertyObject) => {
             var currentValue;
-            if (isToken) {
-                if (objectConfig.object.light.animation) {
-                    currentValue = objectConfig.object.light.animation[customPropertyObject.varName] ?? customPropertyObject.default;
-                } else {
-                    currentValue = customPropertyObject.default;
-                }
+            if (animation) {
+                currentValue = animation[customPropertyObject.varName] ?? customPropertyObject.default;
             } else {
-                if (objectConfig.object.source.animation) {
-                    currentValue = objectConfig.object.source.animation[customPropertyObject.varName] ?? customPropertyObject.default;
-                } else {
-                    currentValue = customPropertyObject.default;
-                }
+                currentValue = customPropertyObject.default;
             }
+
             switch (customPropertyObject.type) {
+                case "speedDescription":
+                    if(currentValue){
+                        var customPropertyEl = $(`<p class="hint community-lighting-custom-property-noanim">${currentValue}</p>`);
+                        $('[name="lightAnimation.speed"]').parent().parent().append(customPropertyEl)
+                    }
+                    break;
+                case "intensityDescription":
+                    if(currentValue){
+                        var customPropertyEl = $(`<p class="hint community-lighting-custom-property-noanim">${currentValue}</p>`);
+                        $('[name="lightAnimation.intensity"]').parent().parent().append(customPropertyEl)
+                    }
+                    break;
                 case "color":
                     var customPropertyEl = $(
                         `<div class="form-group community-lighting-custom-property">
                             <label>${customPropertyObject.title}</label>
                             <input class="color" type="text" name="lightAnimation.${customPropertyObject.varName}" value="${currentValue}">
                             <input type="color" value="${currentValue}" data-edit="lightAnimation.${customPropertyObject.varName}">
+                            ${insertDescription(customPropertyObject)}
                         </div>`)
                     if (animateShow) {
                         customPropertyEl.hide();
@@ -91,12 +112,13 @@ class CLCustomPropertyManager {
                     break;
                 case "range":
                     var customPropertyEl = $(
-                    `<div class="form-group community-lighting-custom-property">
+                        `<div class="form-group community-lighting-custom-property">
                         <label>${customPropertyObject.title}</label>
                         <div class="form-fields">
                             <input type="range" name="lightAnimation.${customPropertyObject.varName}" value="${currentValue}" min="${customPropertyObject.min}" max="${customPropertyObject.max}" step="${customPropertyObject.step}" data-dtype="Number">
                             <span class="range-value">${currentValue}</span>
                         </div>
+                        ${insertDescription(customPropertyObject)}
                     </div>`);
                     if (animateShow) {
                         customPropertyEl.hide();
@@ -107,11 +129,11 @@ class CLCustomPropertyManager {
                     }
                     break;
                 case "checkbox":
-                    var customPropertyEl = 
-                    $(
-                    `<div class="form-group community-lighting-custom-property">
+                    var customPropertyEl = $(
+                        `<div class="form-group community-lighting-custom-property">
                         <label>${customPropertyObject.title}</label>
-                        <input type="checkbox" name="lightAnimation.${customPropertyObject.varName}" data-dtype="Boolean" ${currentValue?"checked":""}>
+                        <input type="checkbox" name="lightAnimation.${customPropertyObject.varName}" data-dtype="Boolean" ${currentValue ? "checked" : ""}>
+                        ${insertDescription(customPropertyObject)}
                     </div>`);
                     if (animateShow) {
                         customPropertyEl.hide();
@@ -120,21 +142,22 @@ class CLCustomPropertyManager {
                     if (animateShow) {
                         customPropertyEl.show('normal');
                     }
-                    
+
                     break;
                 case "select":
                     var options = '';
                     customPropertyObject.options.forEach(option => {
-                        options+=`<option value="${option.value}"${option.value == currentValue?"selected":""}>${option.label}</option>`;
+                        options += `<option value="${option.value}"${option.value == currentValue ? "selected" : ""}>${option.label}</option>`;
                     })
                     var customPropertyEl = $(
-                    `<div class="form-group community-lighting-custom-property">
+                        `<div class="form-group community-lighting-custom-property">
                         <label>${customPropertyObject.title}</label>
                         <div class="form-fields">
                             <select name="lightAnimation.${customPropertyObject.varName}">
                                 ${options}
                             </select>
                         </div>
+                        ${insertDescription(customPropertyObject)}
                     </div>`);
                     if (animateShow) {
                         customPropertyEl.hide();
@@ -143,7 +166,7 @@ class CLCustomPropertyManager {
                     if (animateShow) {
                         customPropertyEl.show('normal');
                     }
-                    
+
                     break;
 
                 default:
@@ -163,6 +186,11 @@ class CLCustomPropertyManager {
     }
 
     static async onRenderTokenOrLightConfig(objectConfig, html, data) {
+        if(objectConfig.object.documentName === 'AmbientLight'){
+            objectConfig.options.closeOnSubmit = game.settings.get(CommunityLighting.moduleName, "closeLightOnSubmit");
+        } else if (objectConfig.object.documentName === 'Token'){
+            objectConfig.options.closeOnSubmit = game.settings.get(CommunityLighting.moduleName, "closeTokenOnSubmit");
+        }
         var animTypeSelector = html.find("[name='lightAnimation.type']"); // Get the animation type selector element
         var customPropertySibling = html.find('[name="lightAnimation.intensity"]').parent().parent(); // Get the div holding the intensity slider so we can add our props after it
         CLCustomPropertyManager.addOptGroups(html);
@@ -172,7 +200,7 @@ class CLCustomPropertyManager {
         CLCustomPropertyManager.removeAllCustomProperties(html);
 
         if (customAnimationProperties && customAnimationProperties.length > 0) {
-            CLCustomPropertyManager.addCustomProperties(objectConfig, customPropertySibling, customAnimationProperties);
+            CLCustomPropertyManager.addCustomProperties(objectConfig.object, customPropertySibling, customAnimationProperties);
         }
 
         // When the type changes, set up any custom properties
@@ -180,14 +208,27 @@ class CLCustomPropertyManager {
             CLCustomPropertyManager.removeAllCustomProperties(html);
             customAnimationProperties = CONFIG.Canvas.lightAnimations[this.value]?.customProperties;
             if (customAnimationProperties && customAnimationProperties.length > 0) {
-                CLCustomPropertyManager.addCustomProperties(objectConfig, customPropertySibling, CONFIG.Canvas.lightAnimations[this.value].customProperties, true);
+                CLCustomPropertyManager.addCustomProperties(objectConfig.object, customPropertySibling, CONFIG.Canvas.lightAnimations[this.value].customProperties, true);
             }
         });
     }
 
-    static onUpdateLightOrToken(scene, object, changes, diff){
-        if(changes.lightAnimation && !diff.loadedProperty && !diff.colorForce){ // Only attempt to save if the lightAnimation prop has changed
-            CLCustomPropertyManager.saveCustomProperties(object, changes);
+    static onPreUpdateLightOrToken(updateData, options, user) {
+        if (options.lightAnimation) {
+            CLCustomPropertyManager.updateCustomProperties(updateData, options.lightAnimation);
+        }
+    }
+
+    static onUpdateLightOrToken(doc, changes, diff, user) {
+        if(!changes?.flags?.CommunityLighting){
+            if(doc.documentName === 'AmbientLight' && game.settings.get(CommunityLighting.moduleName, "closeLightOnSubmit") === false){
+                ui.notifications.notify(game.i18n.localize("COMMUNITYLIGHTING.notif.lightUpdated"));
+            } else if (doc.documentName === 'Token' && game.settings.get(CommunityLighting.moduleName, "closeTokenOnSubmit") === false){
+                ui.notifications.notify(game.i18n.localize("COMMUNITYLIGHTING.notif.tokenUpdated"));
+            }
+        }
+        if (changes?.flags?.CommunityLighting?.customProperties) {
+            CLCustomPropertyManager.saveCustomProperties(doc.data, changes.flags.CommunityLighting.customProperties);
         }
     }